Fix REST API issues
* Fix errors not allowing `PUT` method from the API
* Validate `status`, to allow for status changes when using
PUT (allow changing a `Route`  to `INACTIVE`, re-submitting
`Route`s with `ERROR` state & more)
* Move decision for whether to `commit_*` a `Route` from
`post_save` to `update` since we need to know both the current
`status` of the `Route` and the desired (new) to pick which
`commit` we want
* We need to expose `id`s in all REST API models since those are
needed when creating relationships between those models
* Register `MatchDscp` model (`Route` uses it)
* Add REST API documentation
* When creating / editing / deleting a `Route` from the API an
asynchronous task is issued which uploads the required configuration
on the flowspec device. Since this is asynchronous, the object
must have a status of `PENDING` until this operation is completed.