Fix issue with `request.user` in Celery task
When the user was not a "superuser", `request.user` was being
passed in the route object sent as a Celery job. This led to
an error (only reproducable in one of our instances, don't know
the exact conditions since all packets & versions seemed to be
identical):

    EncodeError: dictionary changed size during iteration

Fix:

* Use the `User` object retrieved from the ORM instead

Also, store the user even if the route is created from an admin.