Fix broken api for non-GET requests
Summary:
- Removes DjangoModelPermissions from the settings template.
DjangoModelPermissions is not needed since the user actions are
validated based on their organization allowed peers. There is no other
permission model/group. DjangoModelPermissions rendered the non-GET
actions unusable until now in the api (only admins could make such
actions) whereas in the views, such actions were allowed
(because the view used django forms instead of api callbacks)
- Downgrades djangorestframework version to solve #1850 drf bug with
pickling and caching. Bug was solved after version 3.0.4 which until now
is not tested, so it's safer to downgrade. More info about the bug:
https://github.com/encode/django-rest-framework/issues/1850
- Fix documentation example for `then` rule
- Adds PENDING in allowed Route statuses since the rules are applied
asynchronously and this means that until applied, the Route will have to
be in PENDING status. This was stated in commit
11171aa200f6baad30e6d0edcc9167585ee31d46 message but was not implemented

Reviewers: #developers, zmousm

Reviewed By: #developers, zmousm

Subscribers: zmousm, #developers

Differential Revision: https://phab.noc.grnet.gr/D5182